user can only crete the post add that id in db

+++ src/main/java/com/vakya/user_service/model/User.java
@@ -27,6 +27,6 @@ public class User {
     private String role;
 
     @ElementCollection
-    private List<Integer> postId;
+    private List<Long> postId;
 }
